Make the webui be started dynamically by etorrent.
This change embeds the webui configuration directly inside etorrent,
rather than relying on a separate configuration file. There is a
regression since we now need to add specifics of binding to a given
port to the code, but otherwise it seem to work.

While here, move log files and private data to the right spots by
utilizing the Erlang/OTP code and application modules.