Fix issue #40: A DivByZero when the Download Rate is 0.
Issue #40 is a simple bug. We divide by zero when starting out since
the download rate at that point is 0. Change the eta code to catch
this and report back an unknown ETA rate.

While here, hoist the formatting code of the ETA rater to its own
function and call this one rather than format it inside the
etorrent:list/0 function.