Narrow down trap_exit from tracker_communication
Normal shutdowns of the tracker will contact the tracker.
Otherwise, we don't care because something is very wrong anyway. Don't
try in that case.

This process *must* trap exits. If it fails because it fails to look
up something, it is an error in the supervision tree order and not in
this process. Basically, we should be allowed to tell the tracker that
we died and is thus in a stopped state.