Remove the operations() type from peer_control/chunk_mgr.
We had this odd way of doing the code here. We passed around the disk IO
operations to be done by the file system all over the code. This is bad
and klaars recent new file IO layer fixed that by lengths. He decided to
do it in the right place and now these IO operation() instructions can
go.