Fix a dialyzer warning (and probably a bug in the DHT subsystem).
Rather than request the Token as a string() type, request it as a
binary() type. This removes a dialyzer warning because
is_valid_token/4 would always fail its membership test. The dialyzer
realizes this by recognizing that lists:member/2 is called with a
term() of type string() on a list of type [binary()]. This will never
ever match.