Fix dialyzer warnings in etorrent_mktorrent. Change API.
The format of bcoded files changed and etorrent_mktorrent were not
brought properly up-to-date before it were pushed. Fix the types and
weed out all dialyzer warnings of the file. While here, tune the CT
suite to recognize that we changed the API.

The API change is to call etorrent_mktorrent:create/3,4 rather than
the redundant etorrent_mktorrent:mktorrent/3.