Fix early incoming connections.
If a connection came in before we were ready for it, it would result
in the untimely death of the etorrent system. The problem is that the
peer_control will want to know the BITFIELD of downloaded and missing
pieces to send to the peer. But this information is unavailable at
that time.

This patch checks if the torrent in question is started properly and
is not in a checking or awaiting state. Otherwise it will gracefully
close down the socket of the incoming peer.