Disconnect from current peers if torrent is private.
In order to comply with BEP 27 Private Torrents, we need to
disconnect from all current peers if the torrent is private.
In addition, according to Azureus wiki, a bittorrent client
that conforms to BEP 27 should behave like classic client
without PEX or DHT. So this is half the private torrent fix.
Should check DHT behaviour under private torrent constraint
in another half of the fix.