Simplify configuration system: Use OTP applications for it.
This change simplifies the configuration system by not using the "new"
gen_server etorrent_config process. Rather it circumvents it and uses
the more direct solution of OTP applications. In principle we can
probably revert the etorrent_config module at some point, but for now
I'll let it pass.

I could imagine that runtime alteration of configuration would be
happy with the idea of the etorrent_config module, but on the other
hand, we could also simply set the things directly in the application.