Fix test cases via rewriting.
We use the newly defined parse_headers/1 call we introduced as a
hack to fix the current eunit test errors. This means we rewrite
parse_msearch_resp/1 to not use mochiweb which will fail. While
here, alter parse_msearch_resp/1 to crash upon error: The code is
internal, and I'd like to outside part to take the crash and quench
the error case if possible.