Postpone startup of the Progress manager.
We now postpone the startup of the Progress manager until we now what
pieces a torrent has and has not. This fixes a problem with negative
left values since it was always set up as if we had downloaded 0
pieces. The negative left values occur due to a responsibility failure
in the code: the progress manager should be responsible, but it is
calculated separately.