Provide a driver for the routing table.
We need to test the routing table without invoking other parts of the system. So we provide a driver for the routing table in order to model its behaviour correctly. The driver process tracks state of the routing table, so we can verify this state transition in a stateful test.