Rework the node insertion code in dht_state.
The code was very convoluted. I initially believed the code to have two code paths,

* One for the case where the node was already a member
* One for when the node was genuinely new

The code then follows the same path, but splits into two paths based on case analysis.
My initial idea was to split this into two separate paths, but it turns out there is only
one path. The first of the above mentioned paths doesn't exist.

When you insert a new node, you should always update its node timer. And you should
always check the bucket timers as well. This is better hoisted out in a function `rt_add/3`
for itself which does the maintenance. This function itself could use some further refactoring,
but this is a start.