Coalesce notifications to the routing table into a `notify/2` call.
This simplifies the interface further. There are three different possible
notifications, which all look kinda the same from the outside. Rather than
having three different functions, build this on one function only and use
variants in the form of atom()'s in order to keep the different kinds of
notifications apart.