Introduce a new dht_state_eqc test framework.
We base our dht_state_eqc framework on EQC Components. The reason we can do this is because
the dht_state code is inherently sequential in nature (most of it at least :), so it is forced to run in
sequence. Hence we don't need to do parallel testing since the observations from the outside is
such that it can never be incorrectly observed.