Split the concerns between `dht_net` and `dht_bt_proto`.
Moving the `handle_query/5` function into `dht_net` makes the protocol encoder be
a real protocol encoder, with no kind of semantic handling at all. This makes it easier to
switch the protocol decoder for another one later on in the project. While here, make ready
for using dht_store as the store for the system.