Adapt the code to a better wire protocol format.
In the wire protocol, we now always supply a 256 ID value from the node. This simplifies a lot of encoding/decoding
paths since there are now always an ID we can use. Also, it simplifies some messages that passed the ID's earlier,
but now they don't have to. Yes, it costs 8 bytes in some messages, but who really cares nowadays. Simpler protocol
parsing should be favored over a couple of bytes. The messages are much better than the competition anyway.