Rewrite the routing code.
We don't really need node timers in this code, so do without them. I think it is easier
to handle node timers at a higher level and setting them when they are needed. They
should never be in a state where we need to cancel them and such.

Range timers can equally be simplified, but the current code base is not doing it.