Cleanup the state code.
Delete functions which has no reason to be public. Also, while here, kill
ping/3 and the UNREACHABLE_TAB for now. A simple model is preferred
over an advanced one with extensions.

Rename a couple of functions which are internal to make them clearer in
what they are trying to do.

Reorder the functions in the module to be systematic and consistent.