EQC Model: Do not shrink to a subtree.
The choice of LETSHRINK looked right, but was wrong. If using LETSHRINK,
we allow the routing table to shrink down to something like the range {0,1}
which then creates all kinds of hell later on because most IDs doesn't fall
within that range. Instead, do the right thing, and don't shrink the subtree.

While here, clean up RT creation.