EQC SUT Fix: Do not use the metric when storing.
For some reason, the old DHT code computed the metric when storing elements in
the routing table. But this leads to numerous errors where the routing table accepts
data which is not valid according to the specification. By removing the distance
calculation, we can simplify insertion by a lot and also simplify many other commands
as a result.

Clean up the remaining code, deleting arguments which are not necessary anymore.

While here, also fix the problem that nodes sometimes were not inserted because we
lacked an ?in_range/3 macro check.