EQC SUT Fix: two grave bugs eliminated.
First bug: We don't know how to correctly call members/2 in dht_routing_table.

Second bug: We don't fold over the right kind of data when folding over ranges.
We should use the data in the routing table, but we are using the data in the
meta-construction which is wrong.

Third bug: Don't ask for ranges late in the process, but do so before insertion
of a new node into the routing table. This is slower, but makes the code do the
functionally right thing.