Reap insert_node/1 for request_success/2
There is no reason to have a way to insert nodes. Rather, once a node is
running successfully, we can just piggyback on this to also insert nodes
into the system.

Ok, this means we have to make the request_success into a call, but this
is needed anyway. But it looks like it is the only way to handle this.