Send tokens with FIND_NODE calls.
If we are searching for a FIND_VALUE target and there is no element in the DHT
already under the ID for which we are searching, we need a way to store an
element under a Node.

This requires we have the Token for that node. The DHT protocol did not send
the token onwards, so it did not work as expected. But this new code sends the
token with the DHT, so it should work as expected.