Handle FIND_NODE correctly in FIND_VALUE searches.
When searching for a value, we may obtain FIND_NODE results. If there are
no entry in the hash table for the target ID, we only have these FIND_NODE
searches as a basis for our work.

Obtain the K closest nodes to the target through successive searches, and
then STORE the value at those.