Make multi-ID storage possible
The bag semantics of the ETS table was used incorrectly, so storing for
instance an IPv4 and IPv6 address on the same ID meant that only one ID
was stored in the DHT.

This patch fixes that problem, by allowing multi-ID storage.

While here, make the EQC model succeed as expected.