Handle the case where a NodeID changes.
Suppose we make a call, and then another NodeID is returned while verifying a node.
The solutions to this question is to handle it by timing out the verifying node and then
trying to insert the new node.

There is no internal message explosion here, since there is an upper bound on how many
messages can flow around until the routing table is up-to-date and then starts draining
these kinds of messages.